added wrist to superstructure

+++ src/main/java/frc/states/SuperStructureStates/BallGetterState.java
@@ -8,7 +8,7 @@ public class BallGetterState extends SuperStructureState {
     }
 
     @Override
-    public double wristPosition() {
+    public int wristPosition() {
         return 6300;
     }
